@@ -44,15 +44,15 @@
     "node-fetch": "^1.3.1",
     "object-assign": "^4.0.1",
     "peerjs": "^0.3.14",
-    "perfect-scrollbar": "^0.6.10",
-    "react": "^0.14.8",
-    "react-addons-css-transition-group": "^0.14.7",
-    "react-dom": "^0.14.8",
+    "perfect-scrollbar": "^0.6.11",
+    "react": "^15.0.2",
+    "react-addons-css-transition-group": "^15.0.2",
+    "react-dom": "^15.0.2",
     "react-foundation-apps": "git+https://github.com/valzav/react-foundation-apps.git",
-    "react-highcharts": "^8.1.0",
-    "react-interpolate-component": "^0.8.0",
+    "react-highcharts": "^8.3.0",
+    "react-interpolate-component": "^0.9.0",
     "react-intl": "^2.0.1",
-    "react-json-inspector": "6.0.0",
+    "react-json-inspector": "^7.0.0",
     "react-motion": "0.3.1",
     "react-notification-system": "^0.2.4",
     "react-popover": "^0.4.4",
