@@ -7,8 +7,8 @@ module.exports = {
         callColor: "#BBBF2B",
         settleColor: "rgba(125, 134, 214, 1)",
         settleFillColor: "rgba(125, 134, 214, 0.5)",
-        positiveColor: "rgba(110, 193, 5, 0.80)",
-        negativeColor: "rgba(225, 66, 74, 0.80)",
+        positiveColor: "rgba(110, 193, 5, 1)",
+        negativeColor: "rgba(225, 66, 74, 1)",
         primaryText: "#e0e0e0",
         lightTextColor: "#fff",
         volumeColor: "#848484",
@@ -30,8 +30,8 @@ module.exports = {
         callColor: "#BBBF2B",
         settleColor: "rgba(125, 134, 214, 1)",
         settleFillColor: "rgba(125, 134, 214, 0.5)",
-        positiveColor: "rgba(110, 193, 5, 0.80)",
-        negativeColor: "rgba(225, 66, 74, 0.80)",
+        positiveColor: "rgba(110, 193, 5, 1)",
+        negativeColor: "rgba(225, 66, 74, 1)",
         primaryText: "#e0e0e0",
         lightTextColor: "#fff",
         volumeColor: "#848484",
@@ -52,8 +52,8 @@ module.exports = {
         callColor: "#BBBF2B",
         settleColor: "rgba(125, 134, 214, 1)",
         settleFillColor: "rgba(125, 134, 214, 0.5)",
-        positiveColor: "rgba(110, 193, 5, 0.80)",
-        negativeColor: "rgba(225, 66, 74, 0.80)",
+        positiveColor: "rgba(110, 193, 5, 1)",
+        negativeColor: "rgba(225, 66, 74, 1)",
         primaryText: "#3F2E55",
         lightTextColor: "#3F2E55",
         volumeColor: "#848484",
