@@ -447,6 +447,7 @@
         settlement_date: "Settlement date",
         asset_reserve: "reserved asset amount",
         trxTypes: {
+            all: "Show all",
             transfer: "Transfer",
             limit_order_create: "Place order",
             limit_order_cancel: "Cancel order",
@@ -922,15 +923,39 @@
         bridge: "Bridge",
         gateway: "Gateway",
         symbol: "Symbol",
+        website: "Visit the website",
         deposit_to: "Deposit To",
-        balance: "Balance",
+        balance: "Current balance",
         generate: "Generate",
-        generate_new: "New address",
+        generate_new: "Get new address",
         deposit: "Deposit",
-        withdraw: "Withdraw",
+        withdraw: "Withdrawal",
+        withdraw_now: "Withdraw now",
         inventory: "Inventory",
         scan_qr: "Scan QR",
         calc: "Calculating",
+        choose_deposit: "Please select the coin you would like to deposit",
+        choose_withdraw: "Please select the coin you would like to withdraw",
+        deposit_to: "Please send your %(asset)s to the address below",
+        asset_to_deposit: "Asset to deposit:",
+        asset_to_receive: "You will receive:",
+        receive_account: "Receiving account:",
+        deposit_summary: "Deposit summary",
+        deposit_inst: "Deposit instructions",
+        recent_deposit: "Recent deposits",
+        recent_withdraw: "Recent withdrawals",
+        withdraw_summary: "Withdrawal summary",
+        withdraw_inst: "Withdrawal instructions",
+        withdraw_to: "Click the button below to open a %(asset)s withdrawal",
+        intermediate: "Intermediate account",
+        asset_to_withdraw: "Asset to withdraw:",
+        support_block: "For support, please contact Blocktrades at:",
+        valid_address: "Please enter a valid %(coin_type)s address",
+        withdraw_coin: "Withdraw %(coin)s (%(symbol)s)",
+        limit: "Limit: %(amount)s %(symbol)s",
+        bridge_text: "A bridge lets you 'shapeshift' an asset into another one, for example BTC to BTS.",
+        gateway_text: "A gateway lets you move to and from Bitshares assets to the real asset, like OPEN.BTC to BTC.",
+        fiat_text: "Fiat deposits and withdrawal are provided by CCEDK. In order to get started you need to complete the registration process. The following conversions are available:",
         transwiser: {
             gateway: "Transwiser",
             visit_weidian: "Visit to deposit",
@@ -951,9 +976,11 @@
         color: "Username color",
         user: "Username",
         done: "Done",
+        reconnect: "Reconnect",
         welcome_user: "Welcome to Bitshares",
+        users: "%(count)s users online",
         welcome: "This is a peer to peer chatroom, with no moderation and no guaranteed unique usernames. It can be disabled in the settings. Message frequency is limited to one per 2 seconds. For a list of available commands, type /help.",
-        disconnected: "You are currently disconnected from the chat, please click below to reconnect.",
+        disconnected: "You are currently not connected to any peers or the central server, please click below to reconnect. If you are unable to connect, this might mean the server is down, please try again later.",
         rate: "Please wait at least 2s before posting a new message"
     }
 };
