@@ -447,6 +447,7 @@
         settlement_date: "Uzlaşma tarihi",
         asset_reserve: "rezerv aktif miktarı",
         trxTypes: {
+            all: "Tümünü göster",
             transfer: "Transfer",
             limit_order_create: "Emir Ver",
             limit_order_cancel: "Emir İptal",
@@ -922,15 +923,39 @@
         bridge: "Köprü",
         gateway: "Kapı",
         symbol: "Sembol",
+        website: "Websitesini ziyaret edin",
         deposit_to: "Yatırılacak Yer",
-        balance: "Bakiye",
+        balance: "Mevcut Bakiye",
         generate: "Yeni Üret",
-        generate_new: "Yeni adres",
+        generate_new: "Yeni adres al",
         deposit: "Yatır",
-        withdraw: "Çek",
+        withdraw: "Para Çekme",
+        withdraw_now: "Şimdi Çek",
         inventory: "Envanter",
         scan_qr: "QR Tara",
         calc: "Hesaplanıyor",
+        choose_deposit: "Lütfen yatırmak istediğiniz coin türünü seçin",
+        choose_withdraw: "Lütfen çekmek istediğiniz coin türünü seçin",
+        deposit_to: "Lütfen %(asset)s adlı aktiflerinizi aşağıdaki adrese gönderin",
+        asset_to_deposit: "Yatırılacak aktif türü:",
+        asset_to_receive: "Alacağınız:",
+        receive_account: "Alacak hesap:",
+        deposit_summary: "Para yatırma özeti",
+        deposit_inst: "Para yatırma talimatları",
+        recent_deposit: "Son para yatırmaları",
+        recent_withdraw: "Son para çekimleri",
+        withdraw_summary: "Para çekim özeti",
+        withdraw_inst: "Para çekim talimatları",
+        withdraw_to: "Aşağıdaki butona tıklayarak %(asset)s adlı aktifler çekme işlemine başlayın",
+        intermediate: "Aracı hesap",
+        asset_to_withdraw: "Çekilecek aktif:",
+        support_block: "Destek için, lütfen Blocktrades'e şuradan ulaşın:",
+        valid_address: "Lütfen geçerli bir  %(coin_type)s adresi giriniz",
+        withdraw_coin: " %(coin)s (%(symbol)s) çekiniz",
+        limit: "Sınır: %(amount)s %(symbol)s",
+        bridge_text: " Bir köprü bir aktifi bir diğerine 'shapeshift' yapmanızı sağlar , mesela BTC'yi BTS'ye.",
+        gateway_text: "Bir kapı BitShares aktiflerinden ve gerçek aktiflere ve geri geçmenizi sağlar, mesela OPEN.BTC 'den  BTC'ye .",
+        fiat_text: "Fiat deposits and withdrawal are provided by CCEDK. In order to get started you need to complete the registration process. The following conversions are available:",
         transwiser: {
             gateway: "Transwiser",
             visit_weidian: "Yatırmak için ziyaret edin",
@@ -951,9 +976,11 @@
         color: "Kullanıcı rengi",
         user: "Kullanıcı adı",
         done: "Tamam",
+        reconnect: "Tekrar bağlan",
         welcome_user: "Bitshares'e Hoşgeldiniz",
-        welcome: "Bu eşten eşe bir sohbet kanalıdır, yöneticisi veya gerçek kullanıcı adları garantisi yoktur. Ayarlar kısmından etkinsizleştirilebilir.Mesaj hızı 2 saniyede bir olarak sınırlıdır",
-        disconnected: "Şu an sohbetle bağlantınız bulunmamaktadır, bağlanmak için lütfen aşağıya tıklayınız.",
+        users: "%(count)s kullanıcı çevrimiçi",
+        welcome: "Bu yöneticisi olmayan ve kullanıcı adlarının gerçek olmama ihtimali olan eşten eşe bir sohbet odasıdır. Ayarlardan kapatılabilir. 2 saniyede 1 mesaj olarak sınırlıdır . Mevcut komutlar için /help yazın..",
+        disconnected: "Şu anda hiçbir eşe yada merkezi sunucuya bağlı değilsiniz, lütfen aşağı tıklayarak tekrar bağlanınız. Bağlanamıyorsanız, bu sunucu down demektir, lütfen sonra tekrar deneyiniz.",
         rate: "Yeni bir mesaj göndermeden önce lütfen en az 2 saniye bekleyin"
     }
 };
