@@ -8,13 +8,25 @@ import {debounce} from "lodash";
 import SettingsActions from "actions/SettingsActions";
 import SettingsStore from "stores/SettingsStore";
 import Peer from "peerjs";
-import Immutable from "immutable";
 import utils from "common/utils";
 import counterpart from "counterpart";
 import LoadingIndicator from "../LoadingIndicator";
 import AccountActions from "actions/AccountActions";
 import TransactionConfirmStore from "stores/TransactionConfirmStore";
 import {FetchChainObjects} from "api/ChainStore";
+
+
+const PROD = true;
+const hostConfig = PROD ? { // Prod config
+    host: 'bitshares.openledger.info',
+    path: '/trollbox',
+    secure: true,
+    port: 443
+} : { // Dev config
+    host: 'localhost',
+    path: '/trollbox',
+    port: 9000
+};
 
 class Comment extends React.Component {
     
@@ -60,19 +72,27 @@ export default class Chat extends React.Component {
     constructor(props) {
         super(props);
 
+        let anonName = "anonymous" + Math.round(10000 * Math.random());
         this.state = {
-            messages: [{user: counterpart.translate("chat.welcome_user"), message: counterpart.translate("chat.welcome")}],
+            messages: [{
+                user: counterpart.translate("chat.welcome_user"),
+                message: counterpart.translate("chat.welcome"),
+                color: "black"
+            }],
             connected: false,
-            showChat: props.viewSettings.get("showChat", false),
+            showChat: props.viewSettings.get("showChat", true),
             myColor: props.viewSettings.get("chatColor", "#904E4E"),
-            userName: props.viewSettings.get("chatUsername", "anonymous"),
+            userName: props.viewSettings.get("chatUsername", anonName),
             shouldScroll: true,
-            loading: true
-        }
+            loading: true,
+            anonName: anonName,
+            docked: props.viewSettings.get("dockedChat", false),
+            hasFetchedHistory: false
+        };
 
         this._peer = null;
 
-        this.connections = Immutable.Map();
+        this.connections = new Map();
         this._myID = null;
 
         this.onChangeColor = debounce(this.onChangeColor, 150);
@@ -101,22 +121,18 @@ export default class Chat extends React.Component {
     }
 
     _connectToServer() {
-        this._peer = new Peer({
-            host: 'bitshares.openledger.info',
-            path: '/trollbox',
-            secure: true,
-            port: 443
-        });
+        this._peer = new Peer(hostConfig);
 
         this._peer.on('open', id => {
             console.log("open, my ID is:", id);
             this._myID = id;
             this.setState({
                 connected: true,
-                loading: false
+                loading: false,
+                open: true
             });
 
-            this._peer.listAllPeers(this._connectToPeers.bind(this));
+            this._peer.listAllPeers(this._connectToPeers.bind(this, true));
         });
 
         this._peer.on('connection', this.onConnection.bind(this));
@@ -124,38 +140,102 @@ export default class Chat extends React.Component {
 
         this._peer.on('error', err => {
             console.log(err);
-            this.setState({
-                connected: false
-            });
+            if (err.message.indexOf("Lost connection to server") !== -1) {
+                this.setState({
+                    open: false
+                });
+            }
 
-            this._peer.reconnect();
+            if (err.message.indexOf("Could not get an ID from the server") !== -1) {
+                this.setState({
+                    open: false,
+                    loading: false,
+                    connected: false
+                });
+            }
+
+            // this._peer.reconnect();
+            // this._connectToServer();
         });
     }
 
-    _connectToPeers(peers) {
+    _broadCastPeers() {
+        let peersArray = this.connections.map((conn, peer) => {
+            return peer;
+        }).toArray();
+
+        this._broadCastMessage({
+            peers: peersArray
+        }, false);
+    }    
+
+    _connectToPeers(broadcast = false, peers) {
+        let shouldUpdate = false;
         if (!Array.isArray(peers)) {
             peers = [peers];
         }
-        console.log("peers:", peers);
+
         // this._peers = Immutable.List(peers);
         peers.forEach(peer => {
-
-            if (peer !== this._myID) {
+            if (peer !== this._myID && !this.connections.has(peer)) {
+                shouldUpdate = true;
                 let conn = this._peer.connect(peer);
+
                 conn.on('data', this._handleMessage);
                 conn.on('close', this.onDisconnect.bind(this, peer));
-                this.connections = this.connections.set(peer, conn);
+                this.connections.set(peer, conn);
             }
         });
-        this.forceUpdate();
+        if (shouldUpdate) {
+            this.forceUpdate();
+        }
+        // if (broadcast) {
+        //     setTimeout(this._broadCastPeers.bind(this), 2000);
+        // }
+
+        // console.log("this.connections:", this.connections);
     }
 
+
+
     _handleMessage(data) {
-        this.state.messages.push(data);
-        if (this.state.messages.length >= 100) {
-            this.state.messages.shift();
+        if ("peers" in data && data.peers.length) {
+            return this._connectToPeers(false, data.peers);
         }
-        this.forceUpdate(this._scrollToBottom.bind(this));
+
+        if ("requestHistory" in data) {
+            return this.sendHistory(this.connections.get(data.requestHistory));
+        }
+
+        if (!this.state.fetchingHistory && data.historyCount && !this.state.hasFetchedHistory) {
+            this.setState({
+                fetchingHistory: true
+            });
+            return this.connections.get(data.id).send({requestHistory: this._myID});
+        }
+
+        if ("history" in data && data.history.length) {
+            this.setState({
+                fetchingHistory: false,
+                hasFetchedHistory: true
+            });
+            
+            data.history.forEach(msg => {
+                this.state.messages.push(msg);
+            });
+            this.forceUpdate();
+            this._scrollToBottom();
+        }
+
+        if ("message" in data && data.user && data.color) {
+            this.state.messages.push(data);
+            if (this.state.messages.length >= 100) {
+                this.state.messages.shift();
+            }
+
+            this.forceUpdate(this._scrollToBottom.bind(this));
+        }
+        
     }
 
     _scrollToBottom() {
@@ -166,6 +246,7 @@ export default class Chat extends React.Component {
 
     _onScroll(e) {
         let {scrollTop, scrollHeight, clientHeight} = this.refs.chatbox;
+
         let shouldScroll = scrollHeight - scrollTop <= clientHeight;
         if (shouldScroll !== this.state.shouldScroll) {
             this.setState({
@@ -174,29 +255,42 @@ export default class Chat extends React.Component {
         }
     }
 
+    sendHistory(c) {
+        c.send({history: this.state.messages.filter((msg) => {return msg.user !== "SYSTEM" && msg.user !== "Welcome to Bitshares"})});
+    }
+
     onConnection(c) {
-        console.log("connection:", c);
-        this.connections = this.connections.set(c.id, c);
+        this.connections.set(c.peer, c);
         c.on('data', this._handleMessage);
-        c.on('close', this.onDisconnect.bind(this, c.id));
+        c.on('close', this.onDisconnect.bind(this, c.peer));
+        setTimeout(() => {c.send({
+            id: this._myID,
+            historyCount: this.state.messages.reduce((value, msg) => {return value + (msg.user !== "SYSTEM" ? 1 : 0)}, 0)})
+        }, 200);
         this.forceUpdate();
     }
 
-    onDisconnect(id) {
-        console.log("disconnected:", id);
-        this.connections = this.connections.delete(id);
+    onDisconnect(peer) {
+        this.connections.get(peer).close();
+        this.connections.delete(peer);
+        this.forceUpdate();
+
+        if (!this.connections.size && !this.state.open) {
+            this.setState({
+                connected: false
+            });
+        }
+
         this.forceUpdate();
     }
 
     onTip(input) {
-        console.log("input:", input);
         Promise.all([
             FetchChainObjects(ChainStore.getAsset, [input.asset]),
             FetchChainObjects(ChainStore.getAccount, [this.props.currentAccount]),
             FetchChainObjects(ChainStore.getAccount, [input.to])
         ])
         .then(objects => {
-            console.log("objects:", objects);
             let asset = objects[0][0];
             let fromAccount = objects[1][0];
             let toAccount = objects[2][0];
@@ -211,7 +305,6 @@ export default class Chat extends React.Component {
                 null,
                 asset.get("id")
             ).then( () => {
-                console.log("transfer.then", this);
                 TransactionConfirmStore.unlisten(this.onTrxIncluded);
                 TransactionConfirmStore.listen(this.onTrxIncluded);
             }).catch( e => {
@@ -223,17 +316,14 @@ export default class Chat extends React.Component {
     }
 
     onTrxIncluded(confirm_store_state) {
-        console.log("confirm_store_state:", confirm_store_state);
         if(confirm_store_state.included && confirm_store_state.broadcast) {
             // this.setState(Transfer.getInitialState());
             TransactionConfirmStore.unlisten(this.onTrxIncluded);
             TransactionConfirmStore.reset();
-            console.log("included");
             this._onTipSuccess();
         } else if (confirm_store_state.closed) {
             TransactionConfirmStore.unlisten(this.onTrxIncluded);
             TransactionConfirmStore.reset();
-            console.log("closed");
         }
     }
 
@@ -245,13 +335,9 @@ export default class Chat extends React.Component {
             color: "#B71A00"
         };
 
-        if (this.connections.size) {
-            this.connections.forEach(c => {
-                c.send(message);
-            });            
-        }
+        // Public and local broadcast
+        this._broadCastMessage(message);
 
-        this._handleMessage(message);
         this.refs.input.value = "";
     }
 
@@ -289,6 +375,7 @@ export default class Chat extends React.Component {
                 "This help: /help"
             ];
 
+            // Only local broadcast
             commands.forEach(command => {
                 this._handleMessage({
                     user: "SYSTEM",
@@ -321,16 +408,25 @@ export default class Chat extends React.Component {
             color: this.state.myColor || "#ffffff"
         };
 
+        // Public and local broadcast
+        this._broadCastMessage(message);
+
+        // Reset input and message timestamp
+        this.refs.input.value = "";
+        this.lastMessage = new Date().getTime();
+    }
+
+    _broadCastMessage(message, local = true) {
+        // Local broadcast
+        if (local) {
+            this._handleMessage(message);
+        }
+        // Public broadcast
         if (this.connections.size) {
             this.connections.forEach(c => {
                 c.send(message);
             });            
         }
-
-        this.refs.input.value = "";
-        this._handleMessage(message);
-
-        this.lastMessage = new Date().getTime();
     }
 
     onToggleChat(e) {
@@ -353,7 +449,6 @@ export default class Chat extends React.Component {
     }
 
     onPickAccount(e) {
-        console.log("onPickAccount:", e);
         this.setState({
             userName: e.target.value
         });
@@ -364,7 +459,6 @@ export default class Chat extends React.Component {
     }
 
     _resetServer() {
-        console.log("_resetServer", this.state.userName);
         this._peer.destroy();
         this._connectToServer();
     }
@@ -387,12 +481,25 @@ export default class Chat extends React.Component {
         this.refs.input.value += userName + ", ";
     }
 
+    _onToggleDock() {
+        this.setState({
+            docked: !this.state.docked
+        });
+
+        SettingsActions.changeViewSetting({
+            dockedChat: !this.state.docked
+        });
+    }
+
     render() {
 
-        let {userName, loading} = this.state;
+        let {userName, loading, docked} = this.state;
 
 
         let messages = this.state.messages.map((msg, index) => {
+            if (!msg.user || !msg.color || !msg.message) {
+                return null;
+            }
             let isMine = msg.user === userName || msg.user === this._myID;
 
             return (
@@ -405,17 +512,19 @@ export default class Chat extends React.Component {
                     isMine={isMine}
                 />
             );
+        }).filter((a) => {
+            return a !== null;
         });
 
         let {showChat, showSettings, connected} = this.state;
 
         let chatStyle = {
-            display: showChat ? "block" : "none",
-            float: "right",
-            height: "35px",
-            margin: "0 .5em",
-            width: "350px",
-            marginRight: "1em"
+            display: !showChat ? "none" : !docked ?"block" : "inherit",
+            float: !docked ? "right" : null,
+            height: !docked ? 35 : null,
+            margin: !docked ? "0 .5em" : null,
+            width: !docked ? 350 : 300,
+            marginRight: !docked ? "1rem" : null
         };
 
         let accountOptions = this.props.linkedAccounts
@@ -430,7 +539,7 @@ export default class Chat extends React.Component {
             return <option key={account} value={account}>{account}</option>;
         }).toArray();
 
-        accountOptions.push(<option key="default" value={"anonymous"}>anonymous</option>)
+        accountOptions.push(<option key="default" value={this.state.anonName}>{this.state.anonName}</option>)
 
 
         let settings = (
@@ -466,21 +575,31 @@ export default class Chat extends React.Component {
         );
 
         return (
-            <div id="chatbox" style={{bottom: this.props.footerVisible ? 36 : 0}}>
+            <div
+                id="chatbox"
+                className={docked ? "chat-docked grid-block" : "chat-floating"}
+                style={{
+                    bottom: this.props.footerVisible && !docked ? 36 : null,
+                    height: !docked ? 35 : null
+                }}
+            >
                 {!showChat ? 
                 <a className="toggle-controlbox" onClick={this.onToggleChat.bind(this)}>
                     <span className="chat-toggle"><Translate content="chat.button" /></span>
                 </a> : null}
                 
-                <div style={chatStyle} className="chatbox">
-                    <div className="flyout grid-block main-content vertical">
+                <div style={chatStyle} className={"chatbox"}>
+                    <div className={"grid-block main-content vertical " + (docked ? "docked" : "flyout")} >
                         <div className="chatbox-title grid-block shrink">
                             <Translate content="chat.title" />
-                            <span>&nbsp;- {this.connections.size + 1} users online</span>
+                            <span>&nbsp;- <Translate content="chat.users" count={this.connections.size + 1} /></span>
+                            <div className="chatbox-pin" onClick={this._onToggleDock.bind(this)}>
+                                {docked ? <Icon className="rotate" name="thumb-tack"/> : <Icon name="thumb-tack"/>}
+                            </div>
                             <div className="chatbox-settings" onClick={this.onToggleSettings.bind(this)}>
                                 <Icon name="cog"/>
                             </div>
-                            <a onClick={this.onToggleChat.bind(this)} className="chatbox-close">&times;</a>
+                            {docked ? null : <a onClick={this.onToggleChat.bind(this)} className="chatbox-close">&times;</a>}
                         </div>
 
                         {loading ? <div><LoadingIndicator /></div> : !connected ? (
@@ -488,19 +607,21 @@ export default class Chat extends React.Component {
                             <div style={{padding: 20}}>
                                 <Translate content="chat.disconnected" />
                                 <div style={{paddingTop: 30}}>
-                                    <div onClick={this._resetServer.bind(this)} className="button">Reconnect</div>
+                                    <div onClick={this._resetServer.bind(this)} className="button"><Translate content="chat.reconnect" /></div>
                                 </div>
                             </div>
                         </div>) : (
-                        <div className="grid-block vertical chatbox-content" ref="chatbox" onScroll={this._onScroll.bind(this)}>
-                            {!showSettings ? <div>{messages}</div> : settings}
+                        <div className="grid-block vertical no-overflow chatbox-content"  onScroll={this._onScroll.bind(this)}>
+                            <div className="grid-content" ref="chatbox">
+                                {!showSettings ? <div>{messages}</div> : settings}
+                            </div>
                         </div>)}
 
                         {!showSettings && connected && !loading ? (
                         <div className="grid-block shrink">
                             <div >
                                 <form onSubmit={this.submitMessage.bind(this)}  className="button-group" style={{marginBottom: 0}}>
-                                    <input style={{marginBottom: 0, width: 300, paddingTop: 5, paddingBottom: 5, backgroundColor: "white", fontSize: 12}} ref="input" type="text" />
+                                    <input style={{marginBottom: 0, width: !docked ? 350 : 300, paddingTop: 5, paddingBottom: 5, backgroundColor: "white", fontSize: 12}} ref="input" type="text" />
                                 </form>
                             </div>
                         </div>) : null}
