@@ -200,8 +200,8 @@ class DepthHighChart extends React.Component {
             let middleValue = (flatAsks[0][0] + flatBids[flatBids.length - 1][0]) / 2;
             let adjustedSpread = spread * power;
             
-            config.xAxis.min = middleValue * 0.1 // middleValue * (this.props.noFrame ? 0.8 : 0.50);
-            config.xAxis.max = middleValue * 1.9; //(this.props.noFrame ? 1.2 : 1.50);
+            config.xAxis.min = middleValue * 0.3 // middleValue * (this.props.noFrame ? 0.8 : 0.50);
+            config.xAxis.max = middleValue * 1.7; //(this.props.noFrame ? 1.2 : 1.50);
          
             // if (adjustedSpread > 0 && adjustedSpread > middleValue) {
             //     config.xAxis.min = Math.max(0, middleValue - 1.5 * adjustedSpread);
