@@ -4,7 +4,7 @@ import {PropTypes} from "react";
 import Highcharts from "highcharts/highstock";
 var ReactHighstock = require("react-highcharts/dist/ReactHighstock");
 import utils from "common/utils";
-import _ from "lodash";
+import {cloneDeep} from "lodash";
 import Translate from "react-translate-component";
 import colors from "assets/colors";
 
@@ -226,7 +226,7 @@ class PriceChart extends React.Component {
         let {priceData, volumeData, quoteSymbol, baseSymbol, base, quote, marketReady,
             indicators, indicatorSettings, latest, bucketSize, theme} = this.props;
 
-        let priceSeriesData = _.cloneDeep(priceData);
+        let priceSeriesData = cloneDeep(priceData);
         let currentIndicator = this.getIndicators(this.props);
 
         let positiveColor = colors[theme].positiveColor;
@@ -306,7 +306,8 @@ class PriceChart extends React.Component {
                     color: negativeColor,
                     lineColor: negativeColor,
                     upColor: positiveColor,
-                    upLineColor: positiveColor
+                    upLineColor: positiveColor,
+                    lineWidth: 2
                 },
                 column: {
                     animation: false,
@@ -413,7 +414,8 @@ class PriceChart extends React.Component {
                         y: 0,
                         zIndex: 99,
                         width: 80
-                    }
+                    },
+                    height: "90%"
                 },
                 {
                     labels: {
@@ -453,7 +455,9 @@ class PriceChart extends React.Component {
                     min: 0,
                     crosshair: {
                         snap: false
-                    }
+                    },
+                    height: "50%",
+                    top: "50%"
             }],
             xAxis: {
                 type: "datetime",
