@@ -125,9 +125,9 @@ class AccountNameInput extends React.Component {
             <div className={class_name}>
                 <label><Translate content="account.name" /></label>
                 <input name="value" type="text" id={this.props.id} ref="input" autoComplete="off"
-                       placeholder={this.props.placeholder} defaultValue={this.props.initial_value}
+                       placeholder={this.props.placeholder}
                        onChange={this.handleChange} onKeyDown={this.onKeyDown}
-                       value={this.state.account_name}/>
+                       value={this.state.account_name || this.props.initial_value}/>
                 <div className="facolor-error">{error}</div>
                 <div className="facolor-warning">{error ? null : warning}</div>
             </div>
