@@ -56,7 +56,7 @@ class SettingsEntry extends React.Component {
                 break;
 
             case "connection":
-                value = connection;
+                value = defaults.indexOf(connection) !== -1 ? connection : defaults[0];
                 options = defaults.map(entry => {
                     let option = entry.translate ? counterpart.translate(`settings.${entry.translate}`) : entry;
                     let key = entry.translate ? entry.translate : entry;
@@ -64,6 +64,7 @@ class SettingsEntry extends React.Component {
                 });
 
                 let defaultConnection = defaults[0];
+
                 let confirmButton = <div style={{padding: "10px"}}><button onClick={this._onConfirm.bind(this)} className="button outline"><Translate content="transfer.confirm" /></button></div>
 
                 optional = (
@@ -104,11 +105,13 @@ class SettingsEntry extends React.Component {
                     value = selected;
                 }
 
+
                 if (defaults) {
-                    options = defaults.map(entry => {
+                    options = defaults.map((entry, index) => {
                         let option = entry.translate ? counterpart.translate(`settings.${entry.translate}`) : entry;
+
                         let key = entry.translate ? entry.translate : entry;
-                        return <option value={option} key={key}>{option}</option>;
+                        return <option value={entry.translate ? entry.translate : entry} key={key}>{option}</option>;
                     })
                 } else {
                     input = <input type="text" defaultValue={value} onBlur={this.props.onChange.bind(this, setting)}/>
@@ -122,7 +125,7 @@ class SettingsEntry extends React.Component {
         if (!value && !options) return null;
 
         if (value && value.translate) {
-            value = counterpart.translate(`settings.${value.translate}`);
+            value = value.translate;
         }
 
         return (
@@ -212,6 +215,14 @@ class Settings extends React.Component {
                 });
                 break;
 
+            case "disableChat":
+                SettingsActions.changeSetting({setting: "disableChat", value: e.target.value === "yes" });
+                break;
+
+            case "showSettles":
+                SettingsActions.changeSetting({setting: "showSettles", value: e.target.value === "yes" });
+                break;
+
             case "unit":
                 let index = findEntry(e.target.value, defaults[setting]);
                 SettingsActions.changeSetting({setting: setting, value: defaults[setting][index]});
@@ -267,6 +278,7 @@ class Settings extends React.Component {
                     ref="ws_modal"
                     apis={defaults["connection"]}
                     api={defaults["connection"].filter(a => {return a === this.state.connection})}
+                    changeConnection={(connection) => {this.setState({connection})}}
                 />
             </div>
         );
