@@ -130,7 +130,7 @@ class AmountSelector extends React.Component {
                     <input 
                            disabled={this.props.disabled}
                            type="text"
-                           value={value}
+                           value={value || ""}
                            placeholder={this.props.placeholder}
                            onChange={this._onChange.bind(this) }
                            tabIndex={this.props.tabIndex}/>
